--- app/controllers/admin_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-# AdminController shows the admin page.
-class AdminController < ApplicationController
-  def index; end
-end
