+++ app/controllers/auth/registrations_controller.rb
@@ -3,6 +3,8 @@
 class Auth::RegistrationsController < Devise::RegistrationsController
   layout "authentication", except: :edit
 
+  skip_before_action :redirect_to_setup
+
   # Re-implemented so the template has some auxiliary instance variables.
   def new
     @have_users = User.any?
