--- app/controllers/updates_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-# UpdatesController handles everything there is on software updates on all the
-# managed nodes.
-class UpdatesController < ApplicationController
-  def index; end
-end
