--- app/views/admin/index.html.slim
@@ -1,2 +1,0 @@
-h1
-  | Admin#index
