+++ app/views/setup/welcome.html.slim
@@ -1,0 +1,29 @@
+= content_for :body_class, "demo-1"
+.alert.alert-info.alert-dismissible role="alert"
+  button.close[type="button" data-dismiss="alert" aria-label="Close"]
+    span[aria-hidden="true"]
+      | &times;
+  i.fa.fa-eye.fa-4x.pull-left aria-hidden="true"
+  span Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.
+h1 Initial CaaSP Configuration
+= form_for :settings, url: setup_path, method: :put do |f|
+  .form-group
+    = f.label :company_name, "Company Name"
+    = f.text_field :company_name, value: Pillar.value(pillar: :company_name), class: "form-control"
+  .form-group
+    = f.label :company_unit, "Company Unit"
+    = f.text_field :company_unit, value: Pillar.value(pillar: :company_unit), class: "form-control"
+  .form-group
+    = f.label :email, "E-Mail"
+    = f.text_field :email, value: Pillar.value(pillar: :email), class: "form-control"
+  .form-group
+    = f.label :country
+    = f.text_field :country, value: Pillar.value(pillar: :country), class: "form-control"
+  .form-group
+    = f.label :state
+    = f.text_field :state, value: Pillar.value(pillar: :state), class: "form-control"
+  .form-group
+    = f.label :city
+    = f.text_field :city, value: Pillar.value(pillar: :city), class: "form-control"
+  .clearfix.steps-container
+    = submit_tag "Next", class: "btn btn-primary pull-right"
