--- app/views/updates/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Updates#index</h1>
-<p>Find me in app/views/updates/index.html.erb</p>
